frontend/pyproject.toml
[project]
name = "pulse-frontend"
version = "0.1.0"
requires-python = "<3.13, >=3.9"
dependencies = [
    "starlette==0.27.0",
    "jinja2==3.1.2",
    "aiofiles==23.2.1",
    "httpx==0.25.2",
    "python-dotenv==1.0.0",
    "uvicorn[standard]==0.24.0",
    "itsdangerous==2.1.2"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
